feat: 댓글 수정과 삭제 권한 부여

+++ Dteam2/diagnosis/views.py
@@ -45,18 +45,22 @@ class CommentListView(views.APIView):
     def get(self,request,first_pk,second_pk,format=None):
         user = get_object_or_404(User, pk=first_pk)
         diagnosis = get_object_or_404(Diagnosis, pk=second_pk)
-        comments= Diag_Comment.objects.filter(originPost=diagnosis, parent=None)
+        comments= Diag_Comment.objects.filter(originPost=diagnosis, user_id=user, parent=None)
         serializer= DiagCommentSerializer(comments, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     
 class AddCommentView(views.APIView):
-    def post(self,request,format=None):
-        serializer = DiagCommentSerializer(data=request.data)
+    def post(self, request, format=None):
+        data = request.data.copy()  
+        data['user_id'] = request.user.id
+        serializer = DiagCommentSerializer(data=data)
+        
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 
 class UpadteCommentView(views.APIView):
@@ -67,11 +71,20 @@ class UpadteCommentView(views.APIView):
      
      def put(self, request, pk, format=None):
         comment = get_object_or_404(Diag_Comment, pk=pk)
-        serializer = DiagCommentSerializer(comment, data=request.data)
+        
+        #댓글 작성자와 접근자의 아이디가 같은지 확인
+        if comment.user_id != request.user:
+            return Response({'error': '이 댓글을 수정할 수 있는 권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
+        
+        data = request.data.copy()
+        data['user_id'] = request.user.id
+        serializer = DiagCommentSerializer(comment, data=data)
+        
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
-        return Response(serializer.errors)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 class DeleteCommentView(views.APIView):
     def get(self,request,pk,format=None):
@@ -81,5 +94,10 @@ class DeleteCommentView(views.APIView):
     
     def delete(self, request, pk, format=None):
         comment = get_object_or_404(Diag_Comment, pk=pk)
+        
+        #댓글 작성자와 접근자의 아이디가 같은지 확인
+        if comment.user_id != request.user:
+            return Response({'error': '이 댓글을 삭제할 수 있는 권한이 없습니다.'}, status=status.HTTP_403_FORBIDDEN)
+        
         comment.delete()
-        return Response({"message":"댓글 삭제 성공"})
+        return Response({"message": "댓글 삭제 성공"})
